# swagger.yaml
swagger: '2.0'
info:
  description: >-
    Projektarbete - Linnea Agetorp, Ellinor Danielsson, Joakim Jonsson, Rickard
    Lundby, Basheer Ghlawanji
  version: 1.0.0
  title: Projektgrupp 1 - NodeJS
  contact:
    name: the team
    email: rickardlundby@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
basePath: /api/
host: 'localhost:3000'
...
paths:
...
  /testimonials:
      get:
        tags:
          - Testimonials
        summary: 'Gets one randomized testimonial'
        description:
          'This will return one testimonial'
        operationId: getTestimonial
        produces:
          - application/json
        responses:
          '200':
            description: 'Successfully got testimonial'
          '500':
            description: 'Failed to get testimonials'
...